# feat: refactor `IsLibReferencedByModule` optimizing source gen for a 66% speed up and 75% less memory usage