### refactor `IsMediatorLibReferencedByTheModule`
Uses a breadth first algorithm to visit modules, keeping track of visited modules and caching known assemblies. Note that caching known assemblies doesn't make a big impact on the benchmark, but will likely help large projects with lots of references.

I've also created optimize_depth #110. This keeps the exisitng algorithm using foreach loops (linq used 10MB of `Func<IModuleSymbol, bool>` and closures), uses an assmebly cache and tracks if a module has been visited at a lower depth. It has similar performance improvements to this PR but will likely be slower with larger projects.

Related: #7

### Benchmarks
#### Original
|  Method |     Mean |    Error |   StdDev |      Gen 0 |    Gen 1 | Allocated |
|-------- |---------:|---------:|---------:|-----------:|---------:|----------:|
| Compile | 81.18 ms | 5.526 ms | 16.29 ms | 14500.0000 | 250.0000 |     23 MB |

#### Changes
|  Method |     Mean |    Error |   StdDev |     Gen 0 |    Gen 1 | Allocated |
|-------- |---------:|---------:|---------:|----------:|---------:|----------:|
| Compile | 26.55 ms | 1.642 ms | 4.842 ms | 2538.4615 | 307.6923 |      6 MB |