@martinothamar 2023-09-18
Thanks, this looks great!
Looks like the build error is due to formatting, if you run `dotnet build` it should autoformat due to `Csharpier.MSBuild` reference from `Directory.Build.Props`