@TimothyMakkison 2023-09-19
> Thanks, this looks great! Looks like the build error is due to formatting

👍 should be fixed now.

> if you run `dotnet build` it should autoformat due to `Csharpier.MSBuild` reference from `Directory.Build.Props`

Had to disable this line, I kept getting a weird dll error everytime I tried to run it.