@martinothamar 2023-09-20
Seems like there are still formatting issues. What kind of errors did you experience? What platform, SDK version etc are you using?