@TimothyMakkison 2023-09-20
> What kind of errors did you experience? 

When I first forked this project I was able to run the test as long as I didn't change any code. When I made changes I would get an error linking to a csproj/config file (can't remember the file or error I thought it had chsarpier.dll in it). Deleting the line would let me run the tests. Unfortunately this change doesnt appear in git.

Running dotnet build in command line doesn't format the code. I'll wipe my folder and try again 👍 

> What platform, SDK version etc are you using?

Rider 2023.1.2, Windows 10, .NET 7