@martinothamar 2023-09-20
> mb did I overwrite your commit?

yeah no worries, I ran `dotnet build Mediator.sln` just to see if I would get errors too. But it seemed to work fine. I see that with your latest push there are some unrelated formatting changes that I don't get on my side. That can happen if we format with different versions of csharpier (which is why I like the MSBuild approach :smile: )