@TimothyMakkison 2023-09-20
Tried using VS and wiping my folder before using `git clone`, neither formatted correctly and I haven't had a build error. 

> That can happen if we format with different versions of csharpier

I was just about to update that 😄