@martinothamar 2023-09-20
yeah but `Directory.Build.props` is using and older version:
```
  <ItemGroup>
    <PackageReference Include="CSharpier.MsBuild" Version="0.16.0">
      <PrivateAssets>all</PrivateAssets>
      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
    </PackageReference>
  </ItemGroup>
```

This package has  an embedded version of the CLI using on 0.16, so both 0.23 and 0.25 will format code that doesn't match. 

Now I don't understand why this package doesn't automatically format everything you build (that's the idea, since it brings pre-build target that invokes the embedded CLI). I've never seen it not work in this way