@martinothamar 2023-09-20
Small clarification - it should autoformat code for debug builds, but produce compile errors for release builds