@martinothamar 2023-09-21
> That would explain a lot 😆. Is there a reason why you are using 0.16.0 over a more recent one?

No reason at all, just haven't updated in a while (since I added it probably). Full time work and a young family leaves little time for OS work unfortunately

Green build, merged, thanks :smile: